backend/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-community
langgraph
tavily-python
langchain-core
python-dotenv
websockets
pydantic
httpx